Added macro-enabled word doc as allowed mime type
Hey, some of us who use this in NFA-LD have been having an issue with uploading docs from a Mac. I took a look and I think its because verbatimized word docs save with the macros enabled on Mac, so I added it as an allowed mime type to hopefully fix that issue. Thanks! - Jflick58