Increase proxy read timeout
Not sure if the loading bottleneck is with the web server loading or local internet connection (based on my tests it’s unlikely to be the latter). If the web server is struggling to respond, Ngnix could be the one sending the timeout.  Up to you, but I would try to increase the timeout to see if it helps.